requirements.txt
numpy>=1.26.0
pandas>=2.1.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.1
faiss-cpu>=1.7.4
streamlit>=1.27.2
nltk>=3.8.1
python-dotenv>=1.0.0
tqdm>=4.66.1